Добавил docx шаблон для согласия об обработке персональных данных
